+++ src/io/ft_printf_utils2.c
@@ -6,16 +6,17 @@
 /*   By: jweber <jweber@student.42Lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/04 13:33:01 by jweber            #+#    #+#             */
-/*   Updated: 2025/02/04 14:04:28 by jweber           ###   ########.fr       */
+/*   Updated: 2025/10/21 17:19:39 by jweber           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_io.h"
 #include <stdarg.h>
+#include <unistd.h>
 
-int	print_percent(va_list *ptr, int *count, int fd)
+ssize_t	print_percent(va_list *ptr, int *count, int fd)
 {
-	int	val;
+	ssize_t	val;
 
 	(void) *ptr;
 	val = ft_putchar_fd('%', fd);
@@ -29,14 +30,14 @@ int	print_percent(va_list *ptr, int *count, int fd)
 	return (0);
 }
 
-int	print_x_lower(va_list *ptr, int *count, int fd)
+ssize_t	print_x_lower(va_list *ptr, int *count, int fd)
 {
-	int	a;
-	int	val;
+	int		a;
+	ssize_t	val;
 
 	(void) count;
 	a = va_arg(*ptr, int);
-	val = ft_putnbr_base_u_fd(a, "0123456789abcdef", fd);
+	val = ft_putnbr_base_u_fd((unsigned int) a, "0123456789abcdef", fd);
 	if (val < 0)
 		return (-1);
 	else
@@ -46,13 +47,13 @@ int	print_x_lower(va_list *ptr, int *count, int fd)
 	}
 }
 
-int	print_x_upper(va_list *ptr, int *count, int fd)
+ssize_t	print_x_upper(va_list *ptr, int *count, int fd)
 {
-	int	a;
-	int	val;
+	int		a;
+	ssize_t	val;
 
 	a = va_arg(*ptr, int);
-	val = ft_putnbr_base_u_fd(a, "0123456789ABCDEF", fd);
+	val = ft_putnbr_base_u_fd((unsigned int)a, "0123456789ABCDEF", fd);
 	if (val < 0)
 		return (-1);
 	else
