+++ src/io/ft_putchar_fd.c
@@ -6,14 +6,15 @@
 /*   By: jweber <jweber@student.42Lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 12:51:54 by jweber            #+#    #+#             */
-/*   Updated: 2025/01/30 16:45:55 by jweber           ###   ########.fr       */
+/*   Updated: 2025/10/21 17:05:34 by jweber           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_io.h"
 #include <unistd.h>
+#include <sys/types.h>
 
-int	ft_putchar_fd(char c, int fd)
+ssize_t	ft_putchar_fd(char c, int fd)
 {
 	return (write(fd, &c, 1));
 }
