+++ src/io/ft_putstr_fd.c
@@ -6,14 +6,15 @@
 /*   By: jweber <jweber@student.42Lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 15:47:24 by jweber            #+#    #+#             */
-/*   Updated: 2025/01/30 16:48:05 by jweber           ###   ########.fr       */
+/*   Updated: 2025/10/21 17:07:55 by jweber           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "ft_io.h"
 #include "ft_string.h"
 #include <unistd.h>
 
-int	ft_putstr_fd(const char *s, int fd)
+ssize_t	ft_putstr_fd(const char *s, int fd)
 {
 	return (write(fd, s, ft_strlen(s)));
 }
